compiler patch for fixing issue 72 (#118)
Fixing compiler issues regarding to the order of symbolic arguments. this fix handles unexpected compiler behavior in the presence of multiple compilations.

With this fix, we can run shapeshifter and swiftkv with arbitrary (SP, TP) combinations. This allows high performance deployments of large models that do not fit in a single GPU.